@@ -1,25 +1,23 @@
 ﻿// Copyright (c) Nate McMaster.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using McMaster.Extensions.CommandLineUtils;
 
-namespace McMaster.DotNet.Serve
-{
-    internal static class ConsoleExtensions
-    {
-        public static void Write(this IConsole console, ConsoleColor color, string message)
-        {
-            console.ForegroundColor = color;
-            console.Write(message);
-            console.ResetColor();
-        }
+namespace McMaster.DotNet.Serve;
 
-        public static void WriteLine(this IConsole console, ConsoleColor color, string message)
-        {
-            console.ForegroundColor = color;
-            console.WriteLine(message);
-            console.ResetColor();
-        }
+internal static class ConsoleExtensions
+{
+    public static void Write(this IConsole console, ConsoleColor color, string message)
+    {
+        console.ForegroundColor = color;
+        console.Write(message);
+        console.ResetColor();
+    }
+
+    public static void WriteLine(this IConsole console, ConsoleColor color, string message)
+    {
+        console.ForegroundColor = color;
+        console.WriteLine(message);
+        console.ResetColor();
     }
 }
