@@ -1,39 +1,35 @@
 // Copyright (c) Nate McMaster.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Generic;
-using System.Threading;
 using Microsoft.AspNetCore.DataProtection.Internal;
 using Microsoft.AspNetCore.DataProtection.KeyManagement;
 using Microsoft.Extensions.Logging.Abstractions;
 using Microsoft.Extensions.Options;
 
-namespace McMaster.DotNet.Serve
+namespace McMaster.DotNet.Serve;
+
+// forwards to XmlKeyManager, but with logging disabled
+internal class KeyManager : IKeyManager
 {
-    // forwards to XmlKeyManager, but with logging disabled
-    internal class KeyManager : IKeyManager
+    private readonly XmlKeyManager _wrapped;
+
+    public KeyManager(IOptions<KeyManagementOptions> options, IActivator activator)
     {
-        private readonly XmlKeyManager _wrapped;
-
-        public KeyManager(IOptions<KeyManagementOptions> options, IActivator activator)
-        {
-            _wrapped = new XmlKeyManager(options, activator, NullLoggerFactory.Instance);
-        }
-
-        public IKey CreateNewKey(DateTimeOffset activationDate, DateTimeOffset expirationDate)
-            => _wrapped.CreateNewKey(activationDate, expirationDate);
-
-        public IReadOnlyCollection<IKey> GetAllKeys()
-            => _wrapped.GetAllKeys();
-
-        public CancellationToken GetCacheExpirationToken()
-            => _wrapped.GetCacheExpirationToken();
-
-        public void RevokeAllKeys(DateTimeOffset revocationDate, string reason = null)
-            => _wrapped.RevokeAllKeys(revocationDate, reason);
-
-        public void RevokeKey(Guid keyId, string reason = null)
-            => _wrapped.RevokeKey(keyId, reason);
+        _wrapped = new XmlKeyManager(options, activator, NullLoggerFactory.Instance);
     }
+
+    public IKey CreateNewKey(DateTimeOffset activationDate, DateTimeOffset expirationDate)
+        => _wrapped.CreateNewKey(activationDate, expirationDate);
+
+    public IReadOnlyCollection<IKey> GetAllKeys()
+        => _wrapped.GetAllKeys();
+
+    public CancellationToken GetCacheExpirationToken()
+        => _wrapped.GetCacheExpirationToken();
+
+    public void RevokeAllKeys(DateTimeOffset revocationDate, string reason = null)
+        => _wrapped.RevokeAllKeys(revocationDate, reason);
+
+    public void RevokeKey(Guid keyId, string reason = null)
+        => _wrapped.RevokeKey(keyId, reason);
 }
