@@ -1,22 +1,20 @@
 // Copyright (c) Nate McMaster.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.Extensions.Options;
 
-namespace McMaster.DotNet.Serve.DefaultExtensions
-{
-    internal static class DefaultExtensionsExtensions
-    {
-        public static IApplicationBuilder UseDefaultExtensions(this IApplicationBuilder app, DefaultExtensionsOptions options)
-        {
-            if (app == null)
-            {
-                throw new ArgumentNullException(nameof(app));
-            }
+namespace McMaster.DotNet.Serve.DefaultExtensions;
 
-            return app.UseMiddleware<DefaultExtensionsMiddleware>(Options.Create(options));
+internal static class DefaultExtensionsExtensions
+{
+    public static IApplicationBuilder UseDefaultExtensions(this IApplicationBuilder app, DefaultExtensionsOptions options)
+    {
+        if (app == null)
+        {
+            throw new ArgumentNullException(nameof(app));
         }
+
+        return app.UseMiddleware<DefaultExtensionsMiddleware>(Options.Create(options));
     }
 }
