@@ -1,11 +1,10 @@
 // Copyright (c) Nate McMaster.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-namespace McMaster.DotNet.Serve.DefaultExtensions
+namespace McMaster.DotNet.Serve.DefaultExtensions;
+
+internal class DefaultExtensionsOptions
 {
-    internal class DefaultExtensionsOptions
-    {
-        public string[] Extensions { get; set; }
-    }
+    public string[] Extensions { get; set; }
 }
 
