@@ -1,14 +1,9 @@
 ﻿// Copyright (c) Nate McMaster.
 // Licensed under the Apache License, Version 2.0. See License.txt in the project root for license information.
 
-using System;
-using System.Collections.Generic;
-using System.Text;
+namespace McMaster.DotNet.Serve.Headers;
 
-namespace McMaster.DotNet.Serve.Headers
+public class HeadersOptions
 {
-    public class HeadersOptions
-    {
-        public IDictionary<string, string> Headers { get; set; }
-    }
+    public IDictionary<string, string> Headers { get; set; }
 }
